+++ src/Zizaco/Mongolid/Cursor/CachableOdmCursor.php
@@ -1,4 +1,4 @@
-<?php namespace Zizaco\Mongolid;
+<?php namespace Zizaco\Mongolid\Cursor;
 
 class CachableOdmCursor implements \Iterator
 {
